+++ data_server/Source/error.c
@@ -1,6 +1,6 @@
 #include "PCR.h"
 
-int Source_Start( char *name, char *type, int size, long debug );
+int Source_Start( char *name, char *type, int size, int debug );
 
 int main()
 {
